api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pytz==2023.3
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
email-validator==2.1.0
